move test case from CsvFunctionsSuite to CsvExpressionsSuite
